# openapi.json
{"openapi":"3.1.0","info":{"title":"文档检索系统","description":"文档检索系统 API","version":"1.0.0"},"paths":{"/health":{"get":{"tags":["健康"],"summary":"健康检查接口","description":"返回 is_success:0k 表示服务正常","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/upload_memory":{"post":{"tags":["基础"],"summary":"文档上传到缓存","description":"返回 is_success:0k 表示提交成功","operationId":"upload_doc_api_upload_memory_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/request_upload_doc"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/save_disk":{"post":{"tags":["基础"],"summary":"缓存保存到硬盘","description":"返回 is_success:0k 表示提交成功 no表示失败","operationId":"save_disk_api_save_disk_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/request_save_disk"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/delete_menory":{"post":{"tags":["基础"],"summary":"删除缓存内区块","description":"返回 is_success:0k 表示提交成功 no表示失败","operationId":"delete_menory_api_delete_menory_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/request_save_disk"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/find":{"post":{"tags":["基础"],"summary":"查询文档","description":"返回 is_success:0k 表示提交成功 no表示失败","operationId":"find_api_find_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/request_find"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"request_find":{"properties":{"prompt":{"type":"string","title":"Prompt","description":"查询内容"},"step":{"type":"string","title":"Step","description":"读取区块数"},"memory_name":{"type":"string","title":"Memory Name","description":"区块名称","default":"group"}},"type":"object","title":"request_find"},"request_save_disk":{"properties":{"memory_name":{"type":"string","title":"Memory Name","description":"区块名称"}},"type":"object","title":"request_save_disk"},"request_upload_doc":{"properties":{"title":{"type":"string","title":"Title","description":"文档标题"},"memory_name":{"type":"string","title":"Memory Name","description":"区块名称"},"content":{"type":"string","title":"Content","description":"文档内容"}},"type":"object","title":"request_upload_doc"}}}}